Dynasty: Split file, also check chapter was moved to volume
